pyproject.toml
[project]
name = "nasa-gcmd"
dynamic = ["version"]
description = "Code to make EnviDat data accessible via NASA Earthdata scrapers."
authors = [
    {name = "Sam Woodcock", email = "samuel.woodcock@wsl.ch"},
]
dependencies = [
    "envidat-utils[dotenv]>=1.2.4",
]
requires-python = ">=3.9"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
version = {from = "__version__.py"}
[tool.pdm.dev-dependencies]
dev = [
    "debugpy>=1.6.2",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.2.0"
version_files = [
    "pyproject.toml:version",
    "__version__.py",
    ".env:APP_VERSION",
]
